+++ src/index.js
@@ -5,12 +5,10 @@ require('dotenv/config');
 const express = require('express');
 const { Server } = require('socket.io');
 const cors = require('cors');
-const MessageService = require('./services/Message.service.js');
-const UserService = require('./services/User.service.js');
-const RoomService = require('./services/Room.service.js');
 const UserController = require('./controllers/User.controller.js');
 const RoomController = require('./controllers/Room.controller.js');
 const MessageController = require('./controllers/Message.controller.js');
+const registerHandlers = require('/handlers/handlers.js');
 const vars = require('./vars.js');
 const app = express();
 
@@ -35,7 +33,7 @@ const io = new Server(server, {
   },
 });
 
-io.on('connection', (socket) => {
+const onConnection = (socket) => {
   if (socket.recovered) {
     // eslint-disable-next-line no-console
     console.log(`User ${socket.id} reconnected`);
@@ -43,86 +41,7 @@ io.on('connection', (socket) => {
     // eslint-disable-next-line no-console
     console.log(`User ${socket.id} connected`);
   }
+  registerHandlers(io, socket);
+};
 
-  socket.on('disconnect', () => {
-    const user = UserService.findUserById(socket.id);
-
-    if (user && user.currentRoom) {
-      const exitMsg = MessageService.buildMsg(`${user.username} left room`, {
-        id: 0,
-        username: 'Admin',
-        currentRoom: user.currentRoom,
-      });
-
-      socket.leave(user.currentRoom.name);
-      socket.broadcast.to(user.currentRoom.name).emit('message', exitMsg);
-    }
-
-    UserService.removeOne(socket.id);
-    io.emit('users', UserService.users);
-
-    // eslint-disable-next-line no-console
-    console.log(`User ${socket.id} disconnected`);
-  });
-
-  socket.on('createUser', (data) => {
-    const user = UserService.buildUser(data);
-
-    UserService.addOne(user);
-    io.emit('users', UserService.users);
-    socket.emit('user', user);
-  });
-
-  socket.on('deleteUser', (data) => {
-    UserService.removeOne(data);
-    io.emit('users', UserService.users);
-  });
-
-  socket.on('joinRoom', (data) => {
-    const user = UserService.updateUserRoom(data.user.id, data.room);
-
-    socket.emit('user', user);
-    io.emit('users', UserService.users);
-
-    const joinMsg = MessageService.buildMsg(`${user.username} joined room`, {
-      id: 0,
-      username: 'Admin',
-      currentRoom: data.room,
-    });
-
-    socket.join(user.currentRoom.name);
-    MessageService.addOne(joinMsg);
-    socket.broadcast.to(user.currentRoom.name).emit('message', joinMsg);
-  });
-
-  socket.on('leaveRoom', (data) => {
-    const exitMsg = MessageService.buildMsg(`${data.username} left room`, {
-      id: 0,
-      username: 'Admin',
-      currentRoom: data.room,
-    });
-
-    socket.leave(data.currentRoom.name);
-    MessageService.addOne(exitMsg);
-    socket.broadcast.to(data.currentRoom.name).emit('message', exitMsg);
-
-    const user = UserService.updateUserRoom(data.id, null);
-
-    socket.emit('user', user);
-    io.emit('users', UserService.users);
-  });
-
-  socket.on('room', async (data) => {
-    const room = RoomService.buildRoom(data);
-
-    RoomService.addOne(room);
-    io.emit('rooms', RoomService.rooms);
-  });
-
-  socket.on('sendMessage', (data) => {
-    const message = MessageService.buildMsg(data.text, data.user);
-
-    MessageService.addOne(message);
-    io.to(message.user.currentRoom.name).emit('message', message);
-  });
-});
+io.on('connection', onConnection);
